Use correct variable for insecure tls connection
The value of `icingadb_database_tls_insecure` is meant to be used in the
IcingaDB config template if the variable is defined.
Prior to this change the condition checked against
`icingadb_database_tls_insecure` while the used value was `icingadb_database_insecure`.

Both cases now refer to `icingadb_database_tls_insecure`.

Fixes #302