Fix exposure of TicketSalt on Satellites / Agents
The 'ticket_salt' in the API feature is now using the constant
'TicketSalt' as its value. This avoids accidental exposure as the actual
secret is kept on a per host basis using constants.
The default for the constant 'TicketSalt' is just an empty string if
unchanged by the user.

Fixes #371